Change post scheme in db, write templates to the handlers

+++ business/data/post/post.go
@@ -40,55 +40,54 @@ func New(log *log.Logger, db *sqlx.DB) Post {
 	}
 }
 
-// Create adds a Post to the database. It returns the created Post with
-// fields like ID and DateCreated populated.
+// Create adds a Post to the database. It returns the created Post with fields like ID and DateCreated populated.
 func (p Post) Create(ctx context.Context, traceID string, claims auth.Claims, np NewPost, now time.Time) (Info, error) {
 	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.post.create")
 	defer span.End()
 
+	// todo: split responses into two (InfoText, InfoLink or smth like that)
 	post := Info{
 		ID:         uuid.New().String(),
-		//Author:     Author{Username: claims.User.Username, ID: claims.User.ID},
+		Author:     Author{Username: claims.User.Username, ID: claims.User.ID},
 		Type:		np.Type,
 		Title:		np.Title,
 		URL:		np.URL,
 		Category:	np.Category,
 		Text:		np.Text,
-		// Votes:		[]Votes{},
-		// Comments:	[]Comment{},
+		Votes:		[]Vote{{claims.User.Username, 1}},
+		Comments:	[]Comment{},
 		DateCreated: now.UTC(),
+		UpvotePercentage: 100,
 	}
 
-	//const q = `
-	//INSERT INTO posts
-	//	(score, views, type, title, url, author, category, votes, comments, date_created, text, post_id)
-	//VALUES
-	//	($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12)`
-	//
-	//p.log.Printf("%s: %s: %s", traceID, "post.Create",
-	//	database.Log(q, post.Score, post.Views, post.Type, post.Title, post.URL, post.Author, post.Category,
-	//		post.Votes, post.Comments, post.DateCreated, post.Text, post.ID),
-	//)
-	//
-	//if _, err := p.db.ExecContext(ctx, q, post.Score, post.Views, post.Type, post.Title, post.URL, post.Author,
-	//	post.Category, post.Votes, post.Comments, post.DateCreated, post.Text, post.ID); err != nil {
-	//	return Info{}, errors.Wrap(err, "inserting product")
-	//}
-
-	const q = `
+	const qPost = `
 	INSERT INTO posts
-		(post_id, score, views, type, title, url, category, text, date_created)
+		(post_id, score, views, type, title, url, category, text, date_created, user_id)
 	VALUES
-		($1, $2, $3, $4, $5, $6, $7, $8, $9)`
+		($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)`
 
 	p.log.Printf("%s: %s: %s", traceID, "post.Create",
-		database.Log(q, post.ID, post.Score, post.Views, post.Type, post.Title, post.URL, post.Category,
-			post.Text, post.DateCreated),
+		database.Log(qPost, post.ID, post.Score, post.Views, post.Type, post.Title, post.URL, post.Category,
+			post.Text, post.DateCreated, post.Author.ID),
 	)
 
-	if _, err := p.db.ExecContext(ctx, q, post.ID, post.Score, post.Views, post.Type, post.Title, post.URL,
-		post.Category, post.Text, post.DateCreated); err != nil {
-		return Info{}, errors.Wrap(err, "inserting product")
+	if _, err := p.db.ExecContext(ctx, qPost, post.ID, post.Score, post.Views, post.Type, post.Title,
+		post.URL, post.Category, post.Text, post.DateCreated, post.Author.ID); err != nil {
+		return Info{}, errors.Wrap(err, "creating post")
+	}
+
+	const qVote = `
+	INSERT INTO votes
+		(post_id, user_id, vote)
+	VALUES
+		($1, $2, $3)`
+
+	p.log.Printf("%s: %s: %s", traceID, "post.Create",
+		database.Log(qVote, post.ID, post.Author.ID, 1),
+	)
+
+	if _, err := p.db.ExecContext(ctx, qVote, post.ID, post.Author.ID, 1); err != nil {
+		return Info{}, errors.Wrap(err, "upvote created post")
 	}
 
 	return post, nil
@@ -103,11 +102,14 @@ func (p Post) Delete(ctx context.Context, traceID string, claims auth.Claims, po
 		return ErrInvalidID
 	}
 
+	post, err := p.QueryByID(ctx, traceID, postID)
+	if err != nil {
+		errors.Wrap(err, "selecting post by ID")
+	}
 
-	//// If you are not an admin. // Todo: check if auth.Claims.Subjects == post.Author.ID
-	//if !claims.Authorized(auth.RoleAdmin) {
-	//	return ErrForbidden
-	//}
+	if claims.User.ID == post.Author.ID {
+		return ErrForbidden
+	}
 
 	const q = `
 	DELETE FROM
@@ -131,16 +133,56 @@ func (p Post) Query(ctx context.Context, traceID string) ([]Info, error) {
 	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.post.query")
 	defer span.End()
 
-	const q = `SELECT * FROM posts`
+	// Get all posts
+	const qPost = `SELECT * FROM posts`
 
 	p.log.Printf("%s: %s: %s", traceID, "post.Query",
-		database.Log(q),
+		database.Log(qPost),
 	)
 
-	var posts []Info
-	if err := p.db.SelectContext(ctx, &posts, q); err != nil {
-		return nil, errors.Wrap(err, "selecting posts")
-	}
+	// todo: uncomment and finish
+	//var posts []PostDB
+	//if err := p.db.SelectContext(ctx, &posts, qPost); err != nil {
+	//	return nil, errors.Wrap(err, "selecting posts")
+	//}
+	//
+	//var info []Info
+	//for _, post := range posts {
+	//	// todo: divide into functions
+	//	// Get posts votes
+	//	const qVotes = `SELECT user_id, vote FROM votes WHERE post_id = $1`
+	//
+	//	p.log.Printf("%s: %s: %s", traceID, "post.Query",
+	//		database.Log(qPost),
+	//	)
+	//
+	//	var votes []Vote
+	//	if err := p.db.SelectContext(ctx, &votes, qVotes, post.ID); err != nil {
+	//		return nil, errors.Wrap(err, "selecting votes")
+	//	}
+	//
+	//	/* todo: think about how to get Authors from the comments
+	//	   (1. one more db requests for every comment or keep two values in db about author)
+	//	   (2. write your own marshalling function)
+	//	*/
+	//	 */
+	//	// Get their comments
+	//	const qComments = `SELECT * FROM comments WHERE post_id = $1`
+	//
+	//	p.log.Printf("%s: %s: %s", traceID, "post.Query",
+	//		database.Log(qPost),
+	//	)
+	//
+	//	var comments []CommentDB
+	//	if err := p.db.SelectContext(ctx, &comments, qVotes, post.ID); err != nil {
+	//		return nil, errors.Wrap(err, "selecting votes")
+	//	}
+	//
+	//	// Get their authors
+	//
+	//}
+
+
 
 	return posts, nil
 }
@@ -198,7 +240,7 @@ func (p Post) QueryByCat(ctx context.Context, traceID string, category string) (
 	return post, nil
 }
 
-// // QueryByUser finds the post identified by a given user. // Todo: think about how to get post by user from DB
+// // QueryByUser finds the post identified by a given user. // Todo: think about how to get post by user from PostDB
 //func (p Post) QueryByUser(ctx context.Context, traceID string, user string) (Info, error) {
 //	ctx, span := trace.SpanFromContext(ctx).Tracer().Start(ctx, "business.data.post.querybyuser")
 //	defer span.End()
